+++ modules/notification/outputs.tf
@@ -1,0 +1,4 @@
+output "s3_bucket_notification_id" {
+  description = "ID of S3 bucket"
+  value       = element(concat(aws_s3_bucket_notification.this.*.id, [""]), 0)
+}